Address Discovery and Server Initiated Connection Migration and NAT Traversal (#3)

+++ src/core/packet_builder.h
@@ -247,7 +247,7 @@ QUIC_INLINE
 BOOLEAN
 QuicPacketBuilderAddFrame(
     _Inout_ QUIC_PACKET_BUILDER* Builder,
-    _In_ uint8_t FrameType,
+    _In_ uint32_t FrameType,
     _In_ BOOLEAN IsAckEliciting
     )
 {
